// src/app/models/instructor.ts
export class Instructor{
    instId: number;
    instAccredNo: string;
    instFirstName: string;
    instLastName: string;
    instMiddleName: string;
    instSuffix: string;
    instRegion: string;
    instCity: string;
    instBrgy: string;
    instLtoPid: string;
    instContactNo: string;
    instDln: string;
    instProvince: string;
    instAddressLine: Date;
    instZipCode: string
}